+++ gapipy/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '2.14.0'
+__version__ = '2.14.1'
 __title__ = 'gapipy'